norm

+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: krahnama <krahnama@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/29 11:00:19 by krahnama          #+#    #+#             */
-/*   Updated: 2026/05/29 12:51:00 by krahnama         ###   ########.fr       */
+/*   Updated: 2026/05/31 00:35:58 by krahnama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	char	*res;
-	int		i;
+	size_t	i;
 
 	if (!s || !f)
 		return (NULL);
@@ -31,7 +31,6 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 	res[i] = '\0';
 	return (res);
 }
-
 /* #include <stdio.h>
 #include <stdlib.h>
